connectivity/requirements.txt
sanic==18.12.0
python-networkmanager==2.1